pyproject.toml
[project]
name = "keyframe"
version = "0.1.0"
description = "調子とリズムを記録するタイピング・画面モニタリングツールなのだ"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mss>=10.1.0",
    "openai>=1.102.0",
    "pillow>=11.3.0",
    "pynput>=1.8.1",
    "pyobjc-framework-cocoa>=11.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.10",
    "ty>=0.0.1a19",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ty]
package = "src"
include-private = false
scripts = { keyframe = "main:main" }
